Вынос покрытия из команды
